role dto

+++ src/main/java/com/example/aquatrack_backend/model/Deuda.java
@@ -1,5 +1,6 @@
 package com.example.aquatrack_backend.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -31,9 +32,9 @@ public class Deuda {
   @DateTimeFormat(pattern = "dd-MM-YYYY HH:mm:ss")
   private LocalDateTime fechaUltimaActualizacion;
 
-  private float monto;
+  private BigDecimal monto;
 
-  private float montoMaximo;
+  private BigDecimal montoMaximo;
 
   @ManyToOne()
   private Domicilio domicilio;
