+++ src/main/java/com/example/aquatrack_backend/model/DeudaPago.java
@@ -1,6 +1,8 @@
 package com.example.aquatrack_backend.model;
 
 
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -22,7 +24,7 @@ public class DeudaPago {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private float montoAdeudadoPago;
+    private BigDecimal montoAdeudadoPago;
 
     @ManyToOne()
     private Deuda deuda;
